Move App into app-root/ and add app header
