+++ src/components/app-root/App/package.json
