+++ src/components/app-root/AppHeader/AppHeader.css
@@ -1,0 +1,7 @@
+.AppHeader {
+    padding: 1em 0;
+}
+
+.AppHeader .logo {
+    margin-right: 1em;
+}
