+++ src/components/app-root/AppHeader/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "AppHeader",
+  "version": "0.0.0",
+  "private": true,
+  "main": "./AppHeader.js"
+}
